Refactor the NuGet v2 badges [resharper powershell chocolatey]
This addresses most of #2921 and refactors the powershell code back into the common nuget v2 service class. It also removes mocked tests of the color logic, replacing them with smaller-bracket tests that accomplish the same thing more concisely.

The ReSharper download badges are broken, which looks like an issue with the upstream API. It’s returning -1 for the download counts. I’ve relaxed the schema temporarily, so only the download badges fail rather than all of them.