Fix [codeclimatecoverage] coverage-percentage route
When I started rewriting the analysis badge I realized I’d broken this part of the route in #3316.

In the long run we might want to deprecate both `coverage-percentage` and `coverage-letter` and move the format argument into a query parameter, though I think for the moment it suffices to get it working again.