Tweak Gitpod config
This disables the comments like https://github.com/badges/shields/pull/3933#issuecomment-526905743 which seem a bit chatty, as the review apps can be accessed from the checks.

This and also turns off builds for branches. For the most part we keep PRs open when branches are at a stage of being ready for useful review, so I think that should be sufficient for us.