Vendor Google fonts in
This patch used `google-font-downloader`[1] to generate a fonts directory
that contains the used Google fonts for shields.io. There are bunch of
reasons for vendoring fonts in, but the first and most important one is
privacy.[2] Another is performance.[3] And the third one is reducing
external dependencies. Keeping those small helps to prevent unexpected
breaking of web interfaces just due to changes or failure of an external
service provider.

[1]: https://github.com/Bloggify/google-font-downloader
[2]: https://github.com/google/fonts/issues/1495
[3]: https://csswizardry.com/2019/05/self-host-your-static-assets/