Fix puppetforge module feedback badges
The forge has stopped to collect user feedback a long time ago.
However, it gained the capability to now run static analysis on modules
and give it a score.  This score has been work in progress for some time
will probably change again in the future, and might include back user
feedback in the future, but as this is moving quite slow, rather than
waiting to unbreak this badge, we can switch to the current API endpoint
that provide this scoring as a drop-in replacement for user feedback.

Since the badge is already labelled "score", this seems a reasonable
move.