fix [DockerVersion] service test (#10690)
* Update docker version tester to check Python instead of Memcached

Memcached tags don't follow semver, might fail depending on which  tag was last added.
Python seems to follow semver for all tags.
fixes #10689

* Update docker version tester to check docker-dev instead of python

its a depricated image and is not expected to change tags
tag is semver

* Update docker version tester to check example-voting-app-vote

deprecated yet by docker, i think we can count on that to be stable