fix [bugzilla] live test case to use new custom base url (#11526)
* fix [bugzilla] live test case to use new custom base url

ecplise seems to be in archive mode and may not work anymore in the future.
this commit fixes live tests failed in our daily tests.

* Update example for baseUrl and bugNumber

eclipse bugzilla is in archive mode, we switched live tests to gnu's gcc.
update the baseUrl example as well as the example bug number to fit both default baseUrl and the new example